refactor: reorganize LDAP, Managed, and OIDC user management views
- Reverted LDAP, Managed, and OIDC vue components back to single file components.
- Created a new UserDetails.vue component to handle user details for LDAP, Managed, and OIDC users.
- Removed userManagementMixin.
- Unified user detail logic and UI, using a userType prop to handle differences between user types.
- Updated API endpoint handling and conditional rendering for managed-user-specific fields.
- Improved maintainability by consolidating similar logic and removing duplication.
