+++ src/views/administration/accessmanagement/LdapUsers.vue
@@ -25,13 +25,15 @@
 
 <script>
 import xssFilters from 'xss-filters';
-import common from '../../../../shared/common';
-import CreateLdapUserModal from '../CreateLdapUserModal';
-import bootstrapTableMixin from '../../../../mixins/bootstrapTableMixin';
-import EventBus from '../../../../shared/eventbus';
-import UserDetails from './UserDetails.vue';
+import common from '../../../shared/common';
+import CreateLdapUserModal from './CreateLdapUserModal.vue';
+import bootstrapTableMixin from '../../../mixins/bootstrapTableMixin';
+import EventBus from '../../../shared/eventbus';
+import UserDetails from '../../components/detail-formatters/UserDetails.vue';
+import i18n from '../../../i18n';
 
 export default {
+  i18n,
   name: 'LdapUsersView',
   props: {
     header: String,
@@ -117,7 +119,12 @@ export default {
         detailFormatter: (index, row) => {
           return this.vueFormatter({
             render: () => (
-              <UserDetails row={row} index={index} rowEvents={this.rowEvents} />
+              <UserDetails
+                userType="ldap"
+                row={row}
+                index={index}
+                rowEvents={this.rowEvents}
+              />
             ),
           });
         },
